优化删除功能

+++ routers/admin/banner.js
@@ -72,10 +72,16 @@ router.get('/delete/:id', async ctx => {
         }*/
 
     let row = data[0]
-    await common.unlink(path.resolve(UPLOAD_DIR,row.src)) //删除本地文件
+    fields.forEach(async({name,type}) => {
+        if (type == 'file'){
+            await common.unlink(path.resolve(UPLOAD_DIR,row.src)) //删除本地文件
+        }
+    })
+
+
     await ctx.db.query(`DELETE FROM ${table} WHERE ID = ?`,[id]); //删除数据库中的文件
 
-    ctx.redirect(`${HTTP_ROOT}/admin/banner`)
+    ctx.redirect(`${HTTP_ROOT}/admin/${page_type}`)
 })
